Temporary fix for the links
I noticed that URLs attached to tweets have the same pattern of image URLs, therefore they were removed from the text by cleanMediaURL function. For the moment, I removed the call to the function, so now texts will contain all the links (including also the links to the attached images, which I would like to remove).